+++ simple_history/manager.py
@@ -1,4 +1,5 @@
 from django.db import models
+from django.db.models import OuterRef, Subquery
 from django.utils import timezone
 
 from simple_history.utils import get_change_reason_from_object
@@ -89,14 +90,17 @@ class HistoryManager(models.Manager):
         model = type(self.model().instance)  # a bit of a hack to get the model
         pk_attr = model._meta.pk.name
         queryset = self.get_queryset().filter(history_date__lte=date)
-        for original_pk in set(queryset.order_by().values_list(pk_attr, flat=True)):
-            changes = queryset.filter(**{pk_attr: original_pk})
-            last_change = changes.latest("history_date")
-            if changes.filter(
-                history_date=last_change.history_date, history_type="-"
-            ).exists():
-                continue
-            yield last_change.instance
+        latest_pk_attr_historic_ids = (
+            queryset.filter(**{pk_attr: OuterRef(pk_attr)})
+            .order_by("-pk")
+            .values("pk")[:1]
+        )
+        latest_historics = queryset.filter(
+            history_id__in=Subquery(latest_pk_attr_historic_ids)
+        )
+        adjusted = latest_historics.exclude(history_type="-").order_by(pk_attr)
+        for historic_item in adjusted:
+            yield historic_item.instance
 
     def bulk_history_create(
         self,