+++ CHANGES.rst
@@ -9,6 +9,7 @@ Unreleased
 - Support ``ignore_conflicts`` in ``bulk_create_with_history`` (gh-733)
 - Use ``asgiref`` when available instead of thread locals (gh-747)
 - Sort imports with isort (gh-751)
+- Queryset ``history.as_of`` speed improvements by calculating in the DB (gh-758)
 
 2.12.0 (2020-10-14)
 -------------------